requirements.txt
python-dotenv>=1.0.0
fastapi==0.104.1
uvicorn==0.24.0
azure-ai-projects
azure-identity>=1.15.0
openai>=1.0.0
tenacity>=8.2.3